Added GET request for incident by id
