+++ sir-client/src/API.tsx
@@ -1,0 +1,47 @@
+import axios from 'axios';
+
+const API_HOST = process.env.REACT_APP_API_HOST ? process.env.REACT_APP_API_HOST : '';
+
+export type Individual = {
+    patient: boolean,
+    familyMember: boolean,
+    adult: boolean,
+    child: boolean,
+    staffMember: boolean,
+    visitor: boolean,
+    volunteer: boolean,
+    other: boolean
+}
+
+export type Patient = {
+    patientName: string,
+    patientSocial: string,
+    patientPhone: string,
+    patientAddress: string,
+}
+
+export type Incident = {
+    incidentDate: string;
+    incidentTime: string;
+    incidentLocation: string;
+    eventType: string;
+    harmOrPotentialHarm: boolean;
+    individualsInvolved: Individual;
+    typeOfEvent: string;
+    effectOnIndividual: string;
+    witnessOneName: string;
+    witnessOnePhone: string;
+    witnessTwoName: string;
+    witnessTwoPhone: string;
+    witnessThreeName: string;
+    witnessThreePhone: string;
+    departmentsInvolved: string;
+    incidentDescription: string;
+    preventativeAction: string;
+    patientInfo: Patient;
+}
+
+export const getIncidentByID = async (id: number): Promise<Incident> => {
+  const response = await axios.get(`${API_HOST}/api/incidents/${id}`);
+  return response.data;
+};
